+++ tests/tests_pytorch/core/test_datamodules.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import pickle
-from argparse import ArgumentParser, Namespace
+from argparse import Namespace
 from dataclasses import dataclass
 from typing import Any, Dict
 from unittest import mock
@@ -130,23 +130,6 @@ class DataDirDataModule(BoringDataModule):
     def __init__(self, data_dir: str):
         super().__init__()
         self.data_dir = data_dir
-
-
-def test_dm_add_argparse_args(tmpdir):
-    parser = ArgumentParser()
-    parser = DataDirDataModule.add_argparse_args(parser)
-    args = parser.parse_args(["--data_dir", str(tmpdir)])
-    assert args.data_dir == str(tmpdir)
-
-
-def test_dm_init_from_argparse_args(tmpdir):
-    parser = ArgumentParser()
-    parser = DataDirDataModule.add_argparse_args(parser)
-    args = parser.parse_args(["--data_dir", str(tmpdir)])
-    dm = DataDirDataModule.from_argparse_args(args)
-    dm.prepare_data()
-    dm.setup("fit")
-    assert dm.data_dir == args.data_dir == str(tmpdir)
 
 
 def test_dm_pickle_after_init():
